Welcome to the start of a beautiful relationship?
This is the first commit of an extracted part of Storyteller's code base,
the HTTP parsing, actioning, and responding mechanism. The main idea behind
this was that there really isn't a good Java HTTP server out there, and we
built one, so why not let other people use the cool thing we built?

At present, because I'm rewriting the whole thing by hand, instead of
copy/pasting (I want to make sure I don't keep Storyteller references in the
source, and there are some optimizations that should be made, that are being
made), this isn't ready to go right now. Only the basic server and request
parser are in here right now, and I don't think the request parse is finished,
yet.

Coming soon, more things, and hopefully a working HTTP server, in Java.