HTTPHandler added in.
The HTTPHandler, what's used by the server to actually cause things to occur,
has been added into the repository.

That said, things technically aren't in a working state right now. The Driver
is the issue, because I put in some references to demo handlers which haven't
yet been created.

Also, javadoc has been added, because I've been trying to make sure that this
is well documented. It's important to me that people be able to understand
what's going on, and use this easily.