Updated code to correct style, more documentation.
Updated most of the source to the correct style (which is currently correct only
in my head, but that's okay, we'll get the style guide up and running at
some point in the future (or is it futre?), but until that point, you're just
going to have to live with me arbitrarily making changes to your code in the
name of "style").

Also fixed up some documentation, and added some TODOs, which should be
addressed at some point. Probably tomorrow, I'm thinking tomorrow...