Switching to HTTPHandlerFactory setup / New Demo
Working on switching away from reflection for determining what kind of
HTTPHandler should be used, and using an HTTPHandlerFactory (which is an
abstract singleton factory, huzzah?) instead.

The idea behind this was discussed. Mostly it provides people with more control,
because the idea is that you just let everything be, and use our methods to make
things work. Well, you lose a good deal of control with the reflection setup
(you can't add extra stuff in your handler constructors is the big thing, which
would mean we wouldn't be able to use this with Storyteller, which is maybe the
whole goal, kinda), which makes it not the best.

As for the demo stuff, I just wanted to make the demo better.