Responding is now in HTTPHandler.
This was done so that we have less extending when it comes time to actually make
things happen. As in, previously if you wanted to adjust how the response should
work, you would need to extend the HTTPServer and HTTPResponse, along with
having all of your handlers. Now you just extend the HTTPHandler and play with
the `writeData` method.

It's a good time.