Started working on adding DELETE requests.
It's basically the same as GET and POST requests, but for DELETE ones.
This is because we've started work on deleting things in Storyteller,
and RESTful principles say we should use DELETE requests. And they're
also a part of the HTTP spec, and I personally think it would be awesome
if we got the whole HTTP spec working.

On a GitHub Issues related note, this is the start of work on issue #11.
I hope that got it referenced, but not closed...