Added the option to use arrays in paths.
You can do this by putting:
`addGET("/path/{Integer... name}");` or
`addGET("/path/{Integer ...} name")` or other variations.

Your functions that are called can look like this:
`public void add(Integer[] numbers);` or
`public void add(Integer... numbers);`
with or without more parameters before the array.

An array must be at the end of the path, or an exception will be thrown
(otherwise, we won't know when to stop adding to that array).