Created a new FileHandler.
This one is much nicer than the other ones. It doesn't have to do anything
special to deal with images. We also set the content type automatically, so we
don't have to do any real work there.

The only kinda bad part of it is line 71 of the FileHander. It's a little hacky
feeling, but it's much better than what we had before.