Implemented what I talked about in #22 before.
Specifically,
https://github.com/dkuntz2/java-httpserver/issues/22#issuecomment-34534402.

There isn't a `{*param}` yet, and there probably won't be, but
everything else is implemented and working.

Specifically the guessing what a handle method wants. Which is pretty
neat. Or, at least that works in theory... (in this case, theory means
there's a test that creates a new handler based on
https://gist.github.com/dkuntz2/8876955, and it creates the stuff just
fine; theory also means after doing some minor changes to our demo it
works too).

More tests are needed, and hopefully it'll be easier to work with now
that the code actually compiles. It was a huge pain having to write all
the code without being able to compile and run it. At least some tests
helped out.