Rewrite main server to be more conventional.
Instead of using just a ton of reflection and things otherwise
considered "not best practices", the main server now follows java
conventions a whole lot more.

What still needs to be done is a rewrite of most of the javadocs,
because basically all of the internals changed, and they're not up to
date.

This commit was maybe a bit too large...