Merge MethodWrapper and Route.
The main reason for doing this is that we needed a separate class to do
all of the magic reflection we previously used to make things work (or
add several hundred lines of code to the HttpHandler, which didn't seem
like a good idea), but because we no longer need to keep track of
separate methods, and have our routes just override the abstract
`handle` method, and pass around real objects, we're able to merge the
layer on top of the handle method with the object holding the handle
method.

Basically, this makes for cleaner, more conventional, and less magical
code. Less magic is always a good thing.